telegrambot/requirements.txt
asgiref==3.6.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
# crontab==1.0.1
Django==4.0
celery==5.3.4
redis==5.0.0
python-decouple==3.8
drf-spectacular==0.26.1

# django-crontab==0.7.1
# python-decouple==3.8
# python-dateutil==2.8.2
dj-database-url==1.2.0
django-filter==22.1
djangorestframework==3.14.0
django-nonrelated-inlines==0.2
djangorestframework-simplejwt==5.2.2
django-reversion==5.0.4
# celery==5.2.7
# redis==4.3.4
# django-celery-beat==2.3.0
python-telegram-bot==13.15
django-jazzmin==2.6.0